refactor: rename model files to better describe purpose of file
